// codeflix-admin/src/types/castMember.ts
export interface Results {
  data: CastMember[];
  links: Links;
  meta: Meta;
}

export interface Result {
  data: CastMember;
  links: Links;
  meta: Meta;
}

export interface CastMember {
  id: string;
  name: string;
  type: number;
  deleted_at: null;
  created_at: string;
  updated_at: string;
}

export interface Links {
  prev: string;
  last: string;
  next: string;
  first: string;
}

export interface Meta {
  to: number;
  from: number;
  path: string;
  total: number;
  perPage: number;
  lastPage: number;
  currentPage: number;
}

export interface CastMemberParams {
  page?: number;
  perPage?: number;
  search?: string;
  isActive?: boolean;
}